requirements.txt
PyExecJS
beautifulsoup4
lxml
openpyxl
furl